fullstack-tiretrack/Back-end-tiretrack-main/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
python-decouple==3.8
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6